+++ milestone/src/main/java/com/app/milestone/service/WithdrawalService.java
@@ -1,9 +1,9 @@
-package com.app.milestone.repository.service;
+package com.app.milestone.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class LikeService {
+public class WithdrawalService {
 }
