+++ milestone/src/test/java/com/app/milestone/service/SchoolServiceTest.java
@@ -1,4 +1,15 @@
 package com.app.milestone.service;
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.Rollback;
+import org.springframework.transaction.annotation.Transactional;
+
+@SpringBootTest
+@Slf4j
+@Transactional
+@Rollback(false)
 public class SchoolServiceTest {
+//    private Schoolservice
 }
